+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # Version 0.0.3 (wip)
 - Planets not building warning includes link to build page
 - Fixed radar page system shortcut logic
-- Research page cosmetics
+- Research summary + copy to cliboard
 - Popup cosmetics
 - Bugfix fleet scan with player without alliance